Allow user to specify which SPI port to use.
This follows the same changes I made for the MAX31855 library, for a different PJRC forum member.

So far I have tested that it builds.

Earlier today I suggested the changes to a different user  and they made the changes and it worked for them.

More details on the thread: https://forum.pjrc.com/threads/62418-second-spi-channel-teensy-4-1?p=275900&viewfull=1#post275900